requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-debug-toolbar==2.0
gunicorn==19.9.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.4
